updated likes

+++ server/services/PostsService.js
@@ -10,21 +10,6 @@ class PostsService {
     return posts
   }
 
-  async createPost(postData) {
-    const post = await dbContext.Posts.create(postData)
-    if (!post) {
-      throw new BadRequest('there is no post')
-    }
-    return post
-  }
-
-  async editPost(postId, postData) {
-    const post = await this.getPostById(postId)
-    post.img = postData.img || post.img
-    await post.save()
-    return post
-  }
-
   async getPostById(postId) {
     const post = await dbContext.Posts.findById(postId)
     if (!post) {
@@ -33,9 +18,45 @@ class PostsService {
     return post
   }
 
-  async removePost(postId) {
+  async createPost(postData) {
+    const post = await dbContext.Posts.create(postData)
+    if (!post) {
+      throw new BadRequest('there is no post')
+    }
+    return post
+  }
+
+  //   TODO we need to fix editPost, it is giving 403 error
+  async editPost(postId, userId, postData) {
     const post = await this.getPostById(postId)
+    if (userId !== post.creatorId.toString()) {
+      throw new Forbidden('You did not make this post')
+    }
+    post.img = postData.img || post.img
+    await post.save()
+    return post
+  }
+
+  async removePost(postId, userId) {
+    const post = await this.getPostById(postId)
+    if (userId !== post.creatorId.toString()) {
+      throw new Forbidden('You did not create this post')
+    }
     await post.delete()
+    return post
+  }
+
+  async editLike(postId, creatorId, value) {
+    const post = await dbContext.Posts.findById(postId)
+    const liked = post.likes.find(l => l.creatorId.toString() === creatorId)
+    if (liked) {
+      liked.value = value
+    } else {
+      post.likes.push({
+        creatorId, value
+      })
+    }
+    await post.save()
     return post
   }
 }
